chore: move bitcoin-address to models to avoid circular dependency, ref LEA-1647
