+++ packages/bitcoin/src/bip322/sign-message-bip322-bitcoinjs.ts
@@ -2,10 +2,9 @@ import { base64 } from '@scure/base';
 import * as btc from '@scure/btc-signer';
 import * as bitcoin from 'bitcoinjs-lib';
 
-import { BitcoinNetworkModes } from '@leather.io/models';
+import { BitcoinAddress, BitcoinNetworkModes } from '@leather.io/models';
 
 import { getBitcoinJsLibNetworkConfigByMode } from '../utils/bitcoin.network';
-import { BitcoinAddress } from '../validation/bitcoin-address';
 import {
   bip322TransactionToSignValues,
   ecPairFromPrivateKey,
