+++ packages/bitcoin/src/coin-selection/coin-selection.ts
@@ -2,10 +2,9 @@ import BigNumber from 'bignumber.js';
 import { validate } from 'bitcoin-address-validation';
 
 import { BTC_P2WPKH_DUST_AMOUNT } from '@leather.io/constants';
-import { Money } from '@leather.io/models';
+import { BitcoinError, Money } from '@leather.io/models';
 import { createMoney, sumMoney } from '@leather.io/utils';
 
-import { BitcoinError } from '../validation/bitcoin-error';
 import { filterUneconomicalUtxos, getSizeInfo, getUtxoTotal } from './coin-selection.utils';
 
 export interface CoinSelectionOutput {
