+++ packages/bitcoin/src/index.ts
@@ -25,8 +25,6 @@ export * from './transactions/generate-unsigned-transaction';
 
 export * from './validation/address-validation';
 export * from './validation/amount-validation';
-export * from './validation/bitcoin-address';
-export * from './validation/bitcoin-error';
 export * from './validation/transaction-validation';
 
 export * from './utils/bitcoin.network';
