+++ packages/bitcoin/src/transactions/generate-unsigned-transaction.ts
@@ -1,6 +1,8 @@
 import { hexToBytes } from '@noble/hashes/utils';
 import * as btc from '@scure/btc-signer';
 
+import { BitcoinError } from '@leather.io/models';
+
 import {
   CoinSelectionRecipient,
   CoinSelectionUtxo,
@@ -9,7 +11,6 @@ import {
 } from '../coin-selection/coin-selection';
 import { BtcSignerDefaultBip32Derivation } from '../signer/bitcoin-signer';
 import { BtcSignerNetwork } from '../utils/bitcoin.network';
-import { BitcoinError } from '../validation/bitcoin-error';
 
 export interface GenerateBitcoinUnsignedTransactionArgs {
   feeRate: number;
