+++ packages/bitcoin/src/utils/lookup-derivation-by-address.ts
@@ -1,10 +1,10 @@
 import { HARDENED_OFFSET, HDKey } from '@scure/bip32';
 
+import { BitcoinAddress } from '@leather.io/models';
 import { createCounter } from '@leather.io/utils';
 
 import { makeTaprootAddressIndexDerivationPath } from '../payments/p2tr-address-gen';
 import { makeNativeSegwitAddressIndexDerivationPath } from '../payments/p2wpkh-address-gen';
-import { BitcoinAddress } from '../validation/bitcoin-address';
 import {
   getNativeSegwitAddress,
   getTaprootAddress,
