+++ packages/bitcoin/src/validation/transaction-validation.ts
@@ -1,11 +1,15 @@
-import { type BitcoinNetworkModes, Money } from '@leather.io/models';
+import {
+  BitcoinAddress,
+  BitcoinError,
+  BitcoinNetworkModes,
+  Money,
+  isValidBitcoinAddress,
+} from '@leather.io/models';
 
 import { calculateMaxSpend } from '../coin-selection/calculate-max-spend';
 import { GetBitcoinFeesArgs } from '../fees/bitcoin-fees';
-import { BitcoinError } from '../validation/bitcoin-error';
-import { isValidBitcoinAddress, isValidBitcoinNetworkAddress } from './address-validation';
+import { isValidBitcoinNetworkAddress } from './address-validation';
 import { isBtcBalanceSufficient, isBtcMinimumSpend } from './amount-validation';
-import { BitcoinAddress } from './bitcoin-address';
 
 interface BitcoinTransaction extends Omit<GetBitcoinFeesArgs, 'recipients'> {
   amount: Money;
