+++ packages/models/package.json
@@ -25,6 +25,7 @@
   "dependencies": {
     "@stacks/stacks-blockchain-api-types": "7.8.2",
     "bignumber.js": "9.1.2",
+    "bitcoin-address-validation": "2.2.3",
     "zod": "3.24.1"
   },
   "devDependencies": {
