+++ packages/models/src/crypto-assets/bitcoin/bitcoin-error.ts
