+++ packages/query/src/bitcoin/ordinals/brc20/brc20-tokens.query.ts
@@ -3,7 +3,8 @@ import { useCallback, useEffect } from 'react';
 import { P2TROut } from '@scure/btc-signer/payment';
 import { useInfiniteQuery } from '@tanstack/react-query';
 
-import type { BitcoinAddress, BitcoinSigner } from '@leather.io/bitcoin';
+import type { BitcoinSigner } from '@leather.io/bitcoin';
+import { BitcoinAddress } from '@leather.io/models';
 import { createNumArrayOfRange } from '@leather.io/utils';
 
 import { useLeatherNetwork } from '../../../leather-query-provider';
